# packages/contracts/openapi/labelhub.yaml
openapi: 3.0.3
info:
  title: LabelHub API
  version: 0.10.0
  description: Contract-first API for the LabelHub modular monolith and AI worker.
servers:
  - url: http://localhost:8080/api
...
security:
  - bearerAuth: []
paths:
...
  /schemas:
    get:
      tags: [Schemas]
      operationId: listSchemas
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: q
          in: query
          schema:
            type: string
        - name: scope
          in: query
          schema:
            $ref: '#/components/schemas/SchemaListScope'
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Owner schemas for the authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedSchemas'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
    post:
      tags: [Schemas]
      operationId: createSchema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchemaRequest'
      responses:
        '201':
          description: Created schema family.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelSchema'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
  /schemas/import:
    post:
      tags: [Schemas]
      operationId: importSchemaTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaImportRequest'
      responses:
        '201':
          description: Imported reusable schema template and its initial published version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaImportResult'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
  /schemas/{schemaId}:
    get:
      tags: [Schemas]
      operationId: getSchema
      parameters:
        - $ref: '#/components/parameters/SchemaId'
      responses:
        '200':
          description: Schema family detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabelSchema'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    delete:
      tags: [Schemas]
      operationId: archiveSchemaTemplate
      description: Archive a library schema template. Task-bound schemas are intentionally not archived through this endpoint.
      parameters:
        - $ref: '#/components/parameters/SchemaId'
      responses:
        '204':
          description: Schema template archived.
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
        '409':
          $ref: '#/components/responses/ErrorStateConflict'
  /schemas/{schemaId}/versions:
    get:
      tags: [Schemas]
      operationId: listSchemaVersions
      parameters:
        - $ref: '#/components/parameters/SchemaId'
      responses:
        '200':
          description: Immutable schema versions for this schema family.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchemaVersion'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    post:
      tags: [Schemas]
      operationId: publishSchemaVersion
      parameters:
        - $ref: '#/components/parameters/SchemaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaVersionRequest'
      responses:
        '201':
          description: Immutable schema version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaVersion'
        '400':
          $ref: '#/components/responses/ErrorBadRequest'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /schemas/{schemaId}/versions/{versionId}:
    get:
      tags: [Schemas]
      operationId: getSchemaVersion
      parameters:
        - $ref: '#/components/parameters/SchemaId'
        - $ref: '#/components/parameters/SchemaVersionId'
      responses:
        '200':
          description: Immutable schema version detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaVersion'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /schemas/{schemaId}/versions/{versionId}/export:
    get:
      tags: [Schemas]
      operationId: exportSchemaVersionPackage
      parameters:
        - $ref: '#/components/parameters/SchemaId'
        - $ref: '#/components/parameters/SchemaVersionId'
      responses:
        '200':
          description: Portable LabelHub schema template package.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaExportPackage'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SchemaId:
      name: schemaId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    SchemaVersionId:
      name: versionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    ErrorBadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorUnauthorized:
      description: Missing, expired, or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorForbidden:
      description: Authenticated user is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorNotFound:
      description: Requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ErrorStateConflict:
      description: Request conflicts with the current business state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/ApiFieldError'
    ApiFieldError:
      description: |
        Field-level error from input validation or business guards.
        `field` uses the OpenAPI camelCase property name, for example `quotaTotal` or `deadlineAt`,
        and must match the request body schema field name.
      type: object
      required: [field, message]
      properties:
        field:
          description: OpenAPI camelCase property name that failed validation.
          type: string
        message:
          description: Human-readable field-level validation or business-guard message.
          type: string
...
    SchemaListScope:
      type: string
      enum: [all, library, task]
    CreateSchemaRequest:
      type: object
      required: [taskId, name]
      properties:
        taskId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    SchemaImportRequest:
      type: object
      required: [name, schemaJson]
      properties:
        name:
          type: string
        description:
          type: string
        schemaJson:
          $ref: '#/components/schemas/SchemaDocument'
    SchemaImportResult:
      type: object
      required: [schema, version]
      properties:
        schema:
          $ref: '#/components/schemas/LabelSchema'
        version:
          $ref: '#/components/schemas/SchemaVersion'
    LabelSchema:
      type: object
      required: [id, name, ownerId, createdAt]
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        ownerId:
          type: integer
          format: int64
        currentVersionId:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time
    PagedSchemas:
      type: object
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LabelSchema'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
    SchemaFieldType:
      type: string
      enum: [text, textarea, number, single_select, multi_select, date, file_upload, rich_text, json_editor, llm_interaction, show_item, nested_object, tab_container]
    SchemaFieldValidation:
      type: object
      properties:
        required:
          type: boolean
        minLength:
          type: integer
          minimum: 0
        maxLength:
          type: integer
          minimum: 0
        min:
          type: number
        max:
          type: number
        pattern:
          type: string
        customFunction:
          type: string
          description: Named LabelHub custom validator. Supported values are nonBlankTrimmed, httpsUrl, and jsonObject.
    SchemaFieldOption:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: string
    SchemaTab:
      type: object
      required: [stableId, label, children]
      properties:
        stableId:
          type: string
        label:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/SchemaField'
    LinkageConditionOp:
      type: string
      enum: [eq, neq, in, notIn, gt, gte, lt, lte, empty, notEmpty]
    LinkageConditionValue:
      description: Runtime condition value. Shape is constrained by op at publish-time validation (scalar for eq/neq/comparison, array for in/notIn, absent for empty/notEmpty). Typed as an open value because OpenAPI cannot express op-dependent value shape.
    LinkageAtomicCondition:
      type: object
      required: [field, op]
      properties:
        field:
          type: string
        op:
          $ref: '#/components/schemas/LinkageConditionOp'
        value:
          $ref: '#/components/schemas/LinkageConditionValue'
    LinkageConditionGroup:
      type: object
      properties:
        allOf:
          type: array
          items:
            $ref: '#/components/schemas/LinkageAtomicCondition'
        anyOf:
          type: array
          items:
            $ref: '#/components/schemas/LinkageAtomicCondition'
    LinkageCondition:
      oneOf:
        - $ref: '#/components/schemas/LinkageAtomicCondition'
        - $ref: '#/components/schemas/LinkageConditionGroup'
    SchemaField:
      type: object
      required: [stableId, label, type]
      properties:
        stableId:
          type: string
        label:
          type: string
        type:
          $ref: '#/components/schemas/SchemaFieldType'
        placeholder:
          type: string
        help:
          type: string
        content:
          type: string
        sourcePath:
          type: string
          description: Dot-separated path into the dataset item payload used by show_item fields.
        aiPrompt:
          type: string
        acceptedFileTypes:
          type: array
          items:
            type: string
        maxFileSizeMb:
          type: integer
          minimum: 1
        validation:
          $ref: '#/components/schemas/SchemaFieldValidation'
        options:
          type: array
          items:
            $ref: '#/components/schemas/SchemaFieldOption'
        children:
          type: array
          items:
            $ref: '#/components/schemas/SchemaField'
        tabs:
          type: array
          items:
            $ref: '#/components/schemas/SchemaTab'
        visibleWhen:
          $ref: '#/components/schemas/LinkageCondition'
        requiredWhen:
          $ref: '#/components/schemas/LinkageCondition'
    SchemaJsonSchemaProperty:
      type: object
      additionalProperties: true
      properties:
        type:
          type: string
        title:
          type: string
        description:
          type: string
        enum:
          type: array
          items: {}
        items:
          $ref: '#/components/schemas/SchemaJsonSchemaProperty'
        properties:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SchemaJsonSchemaProperty'
        required:
          type: array
          items:
            type: string
        minLength:
          type: integer
          minimum: 0
        maxLength:
          type: integer
          minimum: 0
        minimum:
          type: number
        maximum:
          type: number
        pattern:
          type: string
        x-labelhub-customFunction:
          type: string
    SchemaDocument:
      type: object
      description: LabelHub schema document. Legacy published versions use `fields`; new versions use JSON Schema root keywords plus `x-labelhub-fields` for runtime rendering metadata.
      properties:
        fields:
          description: Legacy LabelHub field list retained for immutable historical schema versions.
          type: array
          items:
            $ref: '#/components/schemas/SchemaField'
        x-labelhub-schemaFormatVersion:
          type: integer
          enum: [2]
        $schema:
          type: string
        type:
          type: string
          enum: [object]
        properties:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SchemaJsonSchemaProperty'
        required:
          type: array
          items:
            type: string
        x-labelhub-fields:
          type: array
          items:
            $ref: '#/components/schemas/SchemaField'
        x-labelhub-layout:
          type: object
          additionalProperties: true
    SchemaVersionRequest:
      type: object
      required: [schemaJson]
      properties:
        schemaJson:
          $ref: '#/components/schemas/SchemaDocument'
    SchemaVersion:
      type: object
      required: [id, schemaId, versionNumber, schemaJson, fieldStableIds, contentHash, ownerId]
      properties:
        id:
          type: integer
          format: int64
        schemaId:
          type: integer
          format: int64
        versionNumber:
          type: integer
        schemaJson:
          $ref: '#/components/schemas/SchemaDocument'
        fieldStableIds:
          type: array
          items:
            type: string
        contentHash:
          type: string
        publishedAt:
          type: string
          format: date-time
        ownerId:
          type: integer
          format: int64
    SchemaExportPackage:
      type: object
      required: [packageVersion, schemaId, versionId, versionNumber, name, schemaJson]
      properties:
        packageVersion:
          type: integer
          enum: [1]
        schemaId:
          type: integer
          format: int64
        versionId:
          type: integer
          format: int64
        versionNumber:
          type: integer
        name:
          type: string
        description:
          type: string
        schemaJson:
          $ref: '#/components/schemas/SchemaDocument'
...